Created new layout that is a continous scroll. Removed the router functionality that navigated to separate pages
