+++ src/components/Navigation.jsx
@@ -11,12 +11,11 @@ class Navigation extends Component {
                     </a>
                 </div><br />
                 <nav class="nav_links_div">
-                    <a href="/about" class="link_text">About</a>
-                    <a href="/" class="link_text">My Skills</a>
-                    <a href="/work" class="link_text">Work</a>
-                    <a href="/contact" class="link_text">Contact</a>
+                    <a href="#aboutView" class="link_text">About</a>
+                    <a href="#projectView" class="link_text">Work</a>
+                    <a href="#skillView" class="link_text">My Skills</a>
+                    <a href="#contactView" class="link_text">Contact</a>
                 </nav>
-
             </div>
         )
     }
